pipeline_core/pyproject.toml
[project]
name = "pipeline-core"
version = "0.1.0"
description = "Pure Python data pipeline execution library"
readme = "README.md"
authors = [
    { name = "James Robinson", email = "james.k.robinson@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "duckdb>=0.10",
    "pandas>=2.0",
    "jinja2>=3.0",
    "requests>=2.31",
    "numpy>=1.26",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
